Functional prototype
There are 3 projects (apps) of the solution: ClipboardApp, CopyApp and PasteApp.

ClipboardApp's only purpose is to hold a handle to a memory mapped file (mmf) which stores paths of items selected for copying, in order to keep it "alive".
Memory mapped file is used to bypass writing to the disk and avoid lengthy messages between processes.
Once ClipboardApp obtains the mmf handle, it's put to infinite sleep. This is necessary because we want to do "right-click->copy" on items once, and paste multiple times for the remainder of login session.
At most one intance of ClipboardApp is running at any given moment. This is achived using a named system-wide mutex.

CopyApp starts the ClipboardApp and writes to mmf, the root directory of selected items and filenames of selected items.

PasteApp reads data from mmf and runs one robocopy command per selected directory, plus one robocopy command for all selected files.

Keys must be added in Windows regsitry to extend right-click context menu with robocopyUI copy and paste options. Details of how it's done will be provided in future commit.